+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * `HashSyntax` cop does auto-correction.
 * Allow calls to self to fix name clash with argument [#484](https://github.com/bbatsov/rubocop/pull/484)
+* Renamed `SpaceAroundBraces` to `SpaceAroundBlockBraces`.
+* `SpaceAroundBlockBraces` now has a `NoSpaceBeforeBlockParameters` config option to enforce a style for blocks with parameters like `{|foo| puts }`.
 
 ### Bugs fixed